Removal of unused variables throughout code

+++ Main_User_Workspace/cvs/objects/ccarbon_model/include/carbon_model_utils.h
@@ -125,7 +125,7 @@ public:
     static const int getStartYear();
     static const int getEndYear();
 
-    static const int getConceptualRootKey(const ALandAllocatorItem* const aItem );
+    static const size_t getConceptualRootKey(const ALandAllocatorItem* const aItem );
 
     static double interpYearHelper( const objects::PeriodVector<double>& aPeriodVector,
                                     const unsigned int aYear );
